Initial Django-generated code
You can basically skip over this step if you're generally familiar with
what kinds of files Django adds when you run the `startproject` and
`startapp` commands.

One thing to note is that I always like to take the folder created for
me initally by `django-admin.py startproject <project-name>` and rename
it to `config`. I think this make more sense from a modularity
standpoint. This means that I have two `urls.py` files: one in
`config/`, which serves as a central reference of included URLs, and one
in `gappscmu`, an app I started with `manage.py startapp` that keeps
track of the specific URLs I'll use for this example.