Install the apps
The next few steps will be getting the python-social-auth settings in
place in `settings.py`. As always, the first thing we're going to have
to do is add entries for the apps under the `INSTALLED_APPS` setting.

At this point, you can run `manage.py migrate` if you're using Django
1.7 (which you should be!) or `manage.py makemigrations && manage.py
syncdb` if you're using South + Django 1.6 or below.