More miscellaneous settings
These are some more settings which need to be in place for
python-social-auth to work correctly. There's a good chance you might
want to tweak or remove some of these settings, so check the
documentation if necessary. You're probably safe just including them
though if you don't need to do anything special.