Include the relevant urls
First, we'll need to include the python social auth urls. It's here that
the various URLs for each supported backend are defined. In our case,
we're most interested in the `/login/google-oauth2/` URL that will be
added.

Next, we're going to include the urls for the `gappscmu` app, which at
this point we haven't written. You'll see these in the next step, so
hang on.