Write the templates
In our example, we only needed one template: the home page. If the user
is not logged in, this page shows a link to `/login/google-oauth2/`, the
logic of which is handled by python social auth (because we included it
in `config/urls.py`). If the user is logged in, it grabs and displays
the user's username.